move ssh / console and remote-support to venus-access
More services / logic will need to be moved from the gui to here, so it
is available to all gui's. While at that, things can be made a bit more
flexible / smarter by removing the hardcoded assumptions which are now
in multiple places. In order to do that libudev can be used and a dbus
out service to report the results to the gui's.

As a first step, lets move the essential services for recovery to a
service on its own, so venus-platform can crash without losing access to
the device.