add a setting to (dis)allow mk3 firmware updates
When updating a mk3 with older firmare, the vebus system can shutdown
temporarily. This adds a setting to prevent that from happening
automatically after an Venus update. The setting can be 0, the update
is not allowed. This can be set to 1 by the user and all vebus services
are restarted (and their startup code will then pick the correct version).
Or the value is 2, not applicable if the unit was initially already
programmed with a new version.