Use std::vector instead of QLinkedList
The venus-platform project has a singular dependence on a QLinkedList class that appears to live in the victronenergy/qt project. It's the only external dependence outside of victronenergy/veutil project and if I had to make an educated guess it's some oversight from making the venus-platform project public? Seeing that std::vector is already being used in gui-v2 and veutil it seems to be a natural replacement.

Result: The venus-platform project can be compiled upon git checkout without setting up magic build paths.