introduce the VRM portal and security profile settings
Instead of having fine grained settings for all services, group them together
depending on the intended (local) security and connectivity (VRM portal) wanted.

- When the security profile changed, nginx will be restarted to adopt accordingly.
- Start MQTT depending on the Secure settings

add all the conditions which might require the tunnel

Gui-v2 no longer requires the tunnel for remote console on VRM if it has a known
checksum, however there are other services which do require the tunnel, like
Gui-v1 + remote VNC, Node-Red, SignalK, the EV charger and gui-v2 development.

This explicitly checks for all these cases and if any of them match the tunnel
will be created _unless_ the VRM portal setting is not set to full access. The
actual setup of the tunnel is still handled by venus-access, which also handles
the remote support case. The intention is that remote support / sshd still works
even if the device is in a really bad state.

The VRM portal option replaces the more fine tuned / per service VRM options
with a single one.

https://github.com/victronenergy/venus/issues/1255

Add security profile settings

This is implemented as a simple API. There might arise a need to pass to existing
password along, before allowing the password to be changed. Furthermore unlike
normal settings, it won't return the new value, but only if the password had been
changed succesfully.

- manage the VNC websockify process as well
- allow changing the root password
- invalidate existing logins on password change