Add all the conditions which might require the tunnel
Gui-v2 no longer requires the tunnel for remote console on VRM if it has a known
checksum, however there are other services which do require the tunnel, like
Gui-v1 + remote VNC, Node-Red, SignalK, the EV charger and gui-v2 development.

This explicitly checks for all these cases and if any of them match the tunnel
will be created _unless_ the VRM portal setting is not set to full access. The
actual setup of the tunnel is still handled by venus-access, which also handles
the remote support case. The intention is that remote support / sshd still works
even if the device is in a really bad state.

The VRM portal option replaces the more fine tuned / per service VRM options
with a single one.

https://github.com/victronenergy/venus/issues/1255