test if the brigde registration is faster in Qt
Python3 requests / the mosquitto_bridge_registrator.py script is a really
heavy / slow script for just making a https post. This checks if it isn't
a better idea to just do that in Qt itself.

Retries and notification, updating of the flashmq config and notifying
flashmq are missing at the moment. This is only the post.
In needs qtbase-plugin-qopensslbackend to be installed.