Fix ip address not being updated after reconnecting to ethernet
Sometimes the 'PropertyChanged' signal is missed so the ipv4 properties
are not updated. It may also be that the slot 'servicesChanged' is
called with emtpy properties, but change soon after, before the service
is created and the signals are connected.

The solution is to create the service first with (possibly empty) properties,
so the 'PropertyChanged' signal is connected.
Afterwards, explicitly fetch the properties, and update them in the service.
To avoid having to this everytime the ethernet is connected,
only do it once and keep the ethernet service and its signal connection in place.

https://github.com/victronenergy/venus/issues/1437