cope with connman being restarted
When connman is restarted CmManager::disconnect() will be called and
all services will be freed. Since mEthernetService is not reset it
then points to freed memory and reusing it will cause random crashes.

This typically happens when either is connected with a weak wifi
connection which triggers the restart of connman.