getServiceList: skip removed services
If in a slot connected to serviceRemoved getServiceList() is called
it will crash, since it is still in mServicesOrderList, but no longer
in mServices. So check if service isn't a null pointer in getServiceList.
note: There is currently no code actually doing this, but it is code
waiting for problems to happen.