Do MQTT bridge token registration internally instead of Python
Platform memory use increases by about 2 MB for networking and SSL code,
but it saves 12 MB of process footprint for the Python version.

Time goes from 8000 ms to 110 ms on an idle CCGX.

Includes config generation and retry-mechanism: when there is no
network, it will keep retrying every 60s. This time without a permanent
12 MB process present.

In needs qtbase-plugin-qopensslbackend to be installed.