Inject notification extension
This is an extension of the current injection mechanism.
It accepts two additional arguments, a service name and an ID string.
When this format is used, the notification is NOT set to inactive straight away.
It is the responsibility of the caller to set and clear the active and acknowledged paths.
The ID string is written to the notification /Trigger path and can also be used by the caller to identify the notification.
When the service disappears from dbus, all notifications it has injected will be deactivated, acknowledged and removed.

The call to `/Notifications/Inject` returns a nonnegative integer between 0-19 which indicates the index the notification was injected in.
It returns a negative integer when something has gone wrong.

https://github.com/victronenergy/venus-private/issues/627
https://github.com/victronenergy/venus/issues/1586